# routes/docs/api-spec.yaml
swagger: '2.0'
info:
  description: JavaTeam Project - 하나달
  version: 0.0.1
  title: HanaDal Server
  contact:
    url: https://github.com/HanaDal/HanaDal-Server/issues
host: hanadal-server.herokuapp.com
basePath: /api
...
schemes:
- http
paths:
...
  /qna:
    get:
      tags:
      - QnA
      summary: QnA들 얻어오기
      description: QnA들 얻어오기
      operationId: getQnas
      consumes:
      - application/x-www-form-urlencoded
      - application/json
      produces:
      - application/json
      responses:
        200:
          description: 요청 성공
          schema:
            $ref: '#/definitions/Qnas'
    post:
      tags:
      - QnA
      summary: QnA 작성
      description: QnA 작성하기
      operationId: writeQna
      consumes:
      - application/x-www-form-urlencoded
      - application/json
      produces:
      - application/json
      parameters:
      - in: header
        name: X-Access-Token
        required: true
        type: string
        description: JWT
      - in: formData
        name: title
        required: true
        type: string
        description: 질문 제목
      - in: formData
        name: tags
        required: true
        type: string
        description: 태그들
      - in: formData
        name: content
        required: true
        type: string
        description: 질문 내용
      responses:
        201:
          description: 작성 성공
          schema:
            $ref: '#/definitions/Result'
        403:
          description: 유효하지 않은 JWT
          schema:
            $ref: '#/definitions/Result'
...
  /qna/my-question:
    get:
      tags:
      - QnA
      summary: 내 질문 얻기
      description: JWT로 내 질문 얻기
      operationId: getMyQuestion
      produces:
      - application/json
      parameters:
      - in: header
        name: X-Access-Token
        type: string
        required: true
        description: JWT
      responses:
        200:
          description: 요청 성공
          schema:
            $ref: '#/definitions/Qnas'
        403:
          description: 유효하지 않은 JWt
          schema:
            $ref: '#/definitions/Result'
  /qna/my-answer:
    get:
      tags:
      - QnA
      summary: 내 답변 얻기
      description: JWT로 내 답변 얻기
      operationId: getMyAnswer
      produces:
      - application/json
      parameters:
      - in: header
        name: X-Access-Token
        type: string
        required: true
        description: JWT
      responses:
        200:
          description: 요청 성공
          schema:
            $ref: '#/definitions/Qnas'
        403:
          description: 유효하지 않은 JWt
          schema:
            $ref: '#/definitions/Result'
...
definitions:
  Author:
    type: object
    properties:
      _id:
        type: string
      name:
        type: string
      picture:
        type: string
  Result:
    type: object
    properties:
      result:
        type: string
...
  Qnas:
    type: array
    items:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        tags:
          type: array
          items:
            type: string
        author:
          $ref: '#/definitions/Author'
        answerCount:
          type: integer
...